+++ modules/mem-cache-local/lib/cache.js
@@ -36,6 +36,10 @@ var Cache = module.exports = function (opts) {
 
     var FIXED_HEARTBEAT_INTERVAL = 300000; //5 Mins
 
+    var MAX_SLICE_SIZE = 1024000; // 1 meg is data limit (keep slightly less than 1 meg)
+
+    var SLICE_OVERHEAD = 100; // Braces, Quotes, etc save with every slice
+
     var heartbeat = null;
 
     var self = this;
@@ -95,6 +99,56 @@ var Cache = module.exports = function (opts) {
         }
     }
 
+    function saveSliceBySlice(cacheKey, count, key, dataStr, numSlices, dataSliceSize, duration, cb) {
+        memcached.set(cacheKey + count, dataStr.substr(count * dataSliceSize, dataSliceSize), duration,
+            function (err, result) {
+                if (err || !result) {
+                    self.emit(cacheEvents.ERROR, {key:key, error:err || "unknown"});
+                    return cb({message:'failed', data:{key:key, data:dataStr, duration:duration}, error:err
+                        || "unknown"});
+                }
+                count++;
+                if (count == numSlices) {
+                    self.emit(cacheEvents.NEW, {key:key, duration:duration});
+                    return cb(null, {message:'success', data:result});
+                }
+                else {
+                    return saveSliceBySlice(cacheKey, count, key, dataStr, numSlices, dataSliceSize, duration, cb);
+                }
+            });
+    }
+
+    function fetchSliceBySlice(cacheKey, count, key, slices, numSlices, cb) {
+        if (count == numSlices) {
+            var dataStr = slices.join('');
+            try {
+                var result = JSON.parse(dataStr);
+                self.emit(cacheEvents.HIT, {key:key});
+                return cb(null, {message:'success', data:result});
+            }
+            catch (e) {
+                self.emit(cacheEvents.MISS, {key:key, error:e});
+                return cb({message:'failed - bad JSON', data:{key:key}, error:e});
+            }
+        }
+        else {
+            memcached.get(cacheKey + count, function (err, slice) {
+                if (err) {
+                    self.emit(cacheEvents.MISS, {key:key, error:err});
+                    return cb({message:'failed', data:{key:key}, error:err});
+                }
+                else if (slice) {
+                    slices.push(slice);
+                    return fetchSliceBySlice(cacheKey, count+1, key, slices, numSlices, cb);
+                }
+                else {
+                    self.emit(cacheEvents.MISS, {key:key, error:'unexpected result', result:slice});
+                    return cb({message:'failed', data:{key:key}, error:'unexpected result', result:slice});
+                }
+            });
+        }
+    }
+
     this.put = function (key, data, duration, cb) {
         var cacheKey;
         cb = cb || function (err, result) {
@@ -111,22 +165,51 @@ var Cache = module.exports = function (opts) {
             return cb({message:'No key specified'})
         }
 
+        if(key.length + SLICE_OVERHEAD >= MAX_SLICE_SIZE){
+            self.emit(cacheEvents.ERROR, {error:'key is too large'});
+
+            return cb({message:'key is too large'})
+        }
+
         data = data === undefined || data === null ? '' : data;
 
         duration = _.isNumber(duration) ? duration : DEFAULT_DURATION;
 
         cacheKey = crypto.createHash('md5').update(key).digest('hex');
 
-        memcached.set(cacheKey, {key:key, data:data}, duration, function (err, result) {
-            if (err) {
-                self.emit(cacheEvents.ERROR, {key:key, error:err});
+        var dataStr = JSON.stringify(data);
 
-                return cb({message:'failed', data:{key:key, data:data, duration:duration}, error:err});
-            }
-            self.emit(cacheEvents.NEW, {key:key, duration:duration});
+        if(dataStr.length > MAX_SLICE_SIZE){
+            var dataSliceSize = MAX_SLICE_SIZE - key.length - SLICE_OVERHEAD;
+            var numSlices = Math.ceil(dataStr.length/dataSliceSize);
 
-            return cb(null, {message:'success', data:result});
-        });
+            memcached.set(cacheKey, {key:key, slices:numSlices}, duration, function(err, result){
+                if(err || !result){
+
+                    self.emit(cacheEvents.ERROR, {key:key, error: err || "unknown"});
+
+                    return cb({message:'failed storing slice info for large data',
+                        data:{key:key, data:data, duration:duration}, error: err
+                        || "unknown"});
+                }
+
+                return saveSliceBySlice(cacheKey, 0, key, dataStr, numSlices, dataSliceSize, duration, cb);
+            });
+
+        }
+        else {
+            memcached.set(cacheKey, {key:key, data:dataStr}, duration, function (err, result) {
+                if (err || !result) {
+                    self.emit(cacheEvents.ERROR, {key:key, error: err || "unknown"});
+
+                    return cb({message:'failed', data:{key:key, data:data, duration:duration}, error:err || "unknown"});
+                }
+
+                self.emit(cacheEvents.NEW, {key:key, duration:duration});
+
+                return cb(null, {message:'success', data:result});
+            });
+        }
     }
 
 
@@ -148,24 +231,47 @@ var Cache = module.exports = function (opts) {
 
         cacheKey = crypto.createHash('md5').update(key).digest('hex');
 
-        memcached.get(cacheKey, function (err, result) {
+        memcached.get(cacheKey, function (err, cacheValue) {
             if (err) {
                 self.emit(cacheEvents.MISS, {key:key, error:err});
 
                 return cb({message:'failed', data:{key:key}, error:err});
-            } else if (result && key === result.key) {
-                self.emit(cacheEvents.HIT, {key:key});
-
-                return cb(null, {message:'success', data:result.data});
-            } else {
-                self.emit(cacheEvents.MISS, {key:key, error:'unexpected result', result:result});
-
-                return cb({message:'failed', data:{key:key}, error:'unexpected result', result:result});
+            } else if (cacheValue && cacheValue.data) {
+                var result;
+                try {
+                    result = JSON.parse(cacheValue.data);
+                    if (key === cacheValue.key) {
+                        self.emit(cacheEvents.HIT, {key:key});
+                        return cb(null, {message:'success', data:result});
+                    }
+                    self.emit(cacheEvents.MISS, {key:key, error:'unexpected result - md5 collusion',
+                        result:cacheValue});
+                    return cb({message:'failed', data:{key:key}, error:'unexpected result - md5 collusion',
+                        result:cacheValue});
+                }
+                catch(e){
+                    self.emit(cacheEvents.MISS, {key:key, error:e, result:cacheValue});
+                    return cb({message:'failed', data:{key:key}, error:e, result:cacheValue});
+                }
+            } else if(cacheValue && cacheValue.slices != undefined){ // using undefined because slices is number
+                if (key != cacheValue.key) {
+                    self.emit(cacheEvents.MISS, {key:key, error:'unexpected result - md5 collusion',
+                        result:cacheValue});
+                    return cb({message:'failed', data:{key:key}, error:'unexpected result - md5 collusion',
+                        result:cacheValue});
+                }
+                var slices = [];
+                return fetchSliceBySlice(cacheKey, 0, key, slices, cacheValue.slices, cb);
+            }
+            else {
+                self.emit(cacheEvents.MISS, {key:key, error:'unexpected result', result:cacheValue});
+                return cb({message:'failed', data:{key:key}, error:'unexpected result', result:cacheValue});
             }
 
         });
     }
 }
+
 
 util.inherits(Cache, events.EventEmitter);