+++ CHANGES.md
@@ -1,3 +1,7 @@
+## May 14, 2012
+
+* Support larger than 1mb data in local-mem-cache.
+
 ## May 8, 2012
 
 * Moved maxNestedRequests warning out of the _.each loop in select.js